refactor: move decrypt
